Move corestore.ts to firestore.ts
